Amend retval in declaration and definition to size_t.

+++ code/bmpman/bmpman.cpp
@@ -765,13 +765,13 @@ uint bm_get_signature(int handle) {
 	return bm_bitmaps[bitmapnum].signature;
 }
 
-int bm_get_size(int num) {
+size_t bm_get_size(int num) {
 	int n = num % MAX_BITMAPS;
 
 	Assert(n >= 0);
 	Assert(num == bm_bitmaps[n].handle);
 
-	return bm_bitmaps[n].mem_taken;
+	return (size_t)bm_bitmaps[n].mem_taken;
 }
 
 int bm_get_tcache_type(int num) {
